+++ iti-react/Inputs/PhoneInput.tsx
@@ -1,51 +1,20 @@
 ﻿import * as React from 'react'
-import * as moment from 'moment-timezone'
-import { templateFormatter, templateParser, ReactInput, parseDigit } from 'input-format'
-
+import { templateParser, ReactInput, parseDigit } from 'input-format'
 import { getValidationClass, ValidationFeedback, Validator } from '../Validation'
 import {
     WithValidationInjectedProps,
     withValidation,
     WithValidationProps
 } from '../Validation/WithValidation'
+import {
+    template,
+    formatter,
+    normalizePhoneNumber,
+    lenWithCountryCode,
+    visibleLen
+} from '@interface-technologies/iti-react-core/Util/PhoneNumberUtil'
 
-/* This code should handle a variety of US phone number formats:
- * - with or without country code 1
- * - with or without leading +
- * - less than the required number of digits
- * - with or without punctuation
- */
-
-const visibleLen = 10
-const lenWithCountryCode = visibleLen + 1
-
-function normalizePhoneNumber(phoneNumber: string) {
-    let num = phoneNumber.replace(/[^0-9]/g, '')
-
-    if (num.length > 0 && num[0] !== '1') num = '1' + num
-
-    if (num.length > lenWithCountryCode) {
-        num = num.substring(0, lenWithCountryCode)
-    }
-
-    return num
-}
-
-const template = '(xxx) xxx-xxxx'
 const parser = templateParser(template, parseDigit)
-const formatter = templateFormatter(template)
-
-export function formatPhoneNumber(phoneNumber: string) {
-    if (!phoneNumber) return ''
-
-    const normalized = normalizePhoneNumber(phoneNumber)
-    let noCountry = normalized
-    if (noCountry.length > 0 && noCountry[0] === '1') {
-        noCountry = noCountry.substring(1)
-    }
-
-    return formatter(noCountry).text
-}
 
 interface PhoneInputOwnProps {
     id?: string