+++ iti-react-core/Changelog.md
@@ -18,3 +18,4 @@
     -   Bring in resetPageIfFiltersChanged
     -   Bring in an improved version of preventNonExistentPage
 -   Export Omit type. It can be removed from your .d.ts file and imported from iti-react-core instead
+-   Bring in PhoneNumberUtil from iti-react and export formatPhoneNumber