+++ iti-react/Changelog.md
@@ -452,3 +452,4 @@ const myTab: Tab = ['tabName', 'Tab Label', { className: 'my-tab' }]
 -   Bring in FormCheck
 -   Add replacement for window.alert(): `alert(content: string | React.ReactElement<any>, options?: { title?: string }): Promise<void>`
 -   Add `validated-input` class to the `<div>` that wraps ValidatedInput, .etc
+-   Move formatPhoneNumber to iti-react-core